Reorganize workflow tests
